register book and author in database

+++ src/main/java/tmmscode/literalura/model/ResultsData.java
@@ -11,7 +11,7 @@ public record ResultsData(
         String title,
         List<AuthorsData> authors,
         List<String> languages,
-        int download_count
+        Long download_count
 ) {
     @Override
     public String toString() {
